added AccountService with add\sub operations

+++ src/main/kotlin/ppl/dmitrymix/domclick/interview/entity/Account.kt
@@ -1,10 +1,11 @@
 package ppl.dmitrymix.domclick.interview.entity
 
 import org.springframework.data.annotation.Id
+import java.math.BigDecimal
 
 data class Account(
         @Id
         var id: Long?,
         var userId: Long,
-        var amount: Long
+        var amount: BigDecimal
 )
